+++ tests/rules/test_others.py
@@ -1,6 +1,5 @@
 from functools import partial
 
-from arche.readers.items import Items
 from arche.rules.others import compare_boolean_fields, garbage_symbols
 from arche.rules.result import Level, Outcome
 from conftest import create_named_df, create_result
@@ -72,31 +71,32 @@ def test_compare_boolean_fields(
 
 dirty_inputs = [
     (
-        [
-            {
-                "name": " Blacky Robeburned",
-                "address": "here goes &AMP",
-                "phone": "<h1>144</h1>.sx-prime-pricing-long-row { float: left; }",
-                "rank": 14441,
-            },
-            {
-                "name": "<!--Leprous Jim-->",
-                "address": "Some street",
-                "phone": "1144",
-                "rank": 2_039_857,
-            },
-        ],
+        {
+            "name": [" Blacky Robeburned", "\t<!--Leprous &#9; Jim-->"],
+            "address": [["<br> ", {"v", "&amp;"}], "\xa0house"],
+            "phone": [
+                "<h1>144</h1>.sx-prime-pricing-long-row { float: left; }",
+                {"a": "11"},
+            ],
+            "rank": [141, 2_039_857],
+        },
         {
             Level.ERROR: [
                 (
                     "100.0% (2) items affected",
                     None,
                     {
-                        "100.0% of 'name' values contain ` , -->, <!--`": [0, 1],
-                        "50.0% of 'address' values contain `&AMP`": [0],
+                        "100.0% of 'address' values contain `'&amp;', '<br>', '\\xa0'`": [
+                            0,
+                            1,
+                        ],
+                        "100.0% of 'name' values contain `'\\t', ' ', '&#9;', '-->', '<!--'`": [
+                            0,
+                            1,
+                        ],
                         (
                             "50.0% of 'phone' values contain "
-                            "`.sx-prime-pricing-lo, </h1>, <h1>`"
+                            "`'.sx-prime-pricing-lo', '</h1>', '<h1>'`"
                         ): [0],
                     },
                 )
@@ -116,7 +116,7 @@ dirty_inputs = [
 def test_garbage_symbols(
     raw_items, expected_messages, expected_items_count, expected_err_items_count
 ):
-    assert garbage_symbols(Items.from_array(raw_items)) == create_result(
+    assert garbage_symbols(pd.DataFrame(raw_items)) == create_result(
         "Garbage Symbols",
         expected_messages,
         items_count=expected_items_count,